feat: sign improvements
