+++ src/module/dapp/screen/BrowserScreen/BrowserScreen.tsx
@@ -1,7 +1,7 @@
-import WebView from "react-native-webview";
 import { BrowserScreenRoot } from "./BrowserScreen.styles";
 import { useBrowserScreen } from "./hooks/useBrowserScreen";
 import BrowserScreenHeader from "./BrowserScreenHeader/BrowserScreenHeader";
+import DAppWebView from "module/signer/containers/DAppWebView/DAppWebView";
 
 const BrowserScreen = (): JSX.Element => {
     const { headerProps, webviewProps } = useBrowserScreen();
@@ -9,7 +9,7 @@ const BrowserScreen = (): JSX.Element => {
     return (
         <BrowserScreenRoot>
             <BrowserScreenHeader {...headerProps} />
-            <WebView {...webviewProps} />
+            <DAppWebView {...webviewProps} />
         </BrowserScreenRoot>
     );
 };
