--- src/module/signer/components/display/DApp/DApp.types.ts
@@ -1,7 +1,0 @@
-import { DApp } from "module/signer/types";
-
-export interface DAppProps {
-    dapp: DApp;
-    connected?: boolean;
-    loading?: boolean;
-}
