--- src/module/signer/components/feedback/DisconnectableDApp/DisconnectableDApp.types.ts
@@ -1,5 +1,0 @@
-import { DApp } from "module/signer/types";
-
-export interface DisconnectableDAppProps {
-    dapp: DApp;
-}
