+++ src/module/signer/containers/DAppWebViewModal/DAppWebViewModal.styles.ts
@@ -1,6 +1,5 @@
 import styled from "@peersyst/react-native-styled";
 import CardNavigatorModal from "module/common/component/navigation/CardNavigatorModal/CardNavigatorModal";
-import WebView from "react-native-webview";
 
 export const DAppWebViewModalRoot = styled(CardNavigatorModal)(() => ({
     height: "100%",
@@ -8,8 +7,4 @@ export const DAppWebViewModalRoot = styled(CardNavigatorModal)(() => ({
         padding: 0,
         paddingBottom: 0,
     },
-}));
-
-export const DAppWebView = styled(WebView)(() => ({
-    backgroundColor: "transparent",
 }));
